refactor injector tokenizer
