+++ fusedrug/data/tokenizer/modulartokenizer/modular_tokenizer.py
@@ -45,7 +45,7 @@ class ModularTokenizer(transformers.PreTrainedTokenizerFast):
         self,
         tokenizers_info: Union[List, omegaconf.listconfig.ListConfig],
         load_adjusted_jsons: Optional[bool] = False,
-        special_tokens_dict: Optional[Dict] = None,
+        special_tokens_list: Optional[list] = None,
         max_possible_token_id: Optional[int] = None,
         max_special_token_id: Optional[int] = None,
         **kwargs: Any,
@@ -68,8 +68,7 @@ class ModularTokenizer(transformers.PreTrainedTokenizerFast):
             load_adjusted_jsons (Optional[bool], optional): Whether to load json files created by ModularTokenizer (True),
                 or to adjust the indices of given non-modular jsons (False). This should not ordinarily be set to False, as loading
                 from modular jsons is best done through the load_from_jsons method. Defaults to False.
-            special_tokens_dict (Optional[Dict], optional): A dictionary of special tokens that should be common among all tokenizers, with keys
-                from ["bos_token", "eos_token", "unk_token", "sep_token", "pad_token", "cls_token", "mask_token"]
+            special_tokens_list (Optional[list], optional): A list of special tokens that should be common among all tokenizers
             max_possible_token_id (Optional[int], optional): An upper limit to a token ID. When IDs of tokens added to modular tokenizer
                 go above this, an exception is thrown. Defaults to None (i.e. no limit is set).
             max_special_token_id (Optional[int], optional): An upper limit to special token ID. Special tokens are shared between all sub-tokenizers.
@@ -92,17 +91,14 @@ class ModularTokenizer(transformers.PreTrainedTokenizerFast):
         self.tokenizers_info = ModularTokenizer.cfg_list_2_dict(
             copy.deepcopy(tokenizers_info_list)
         )
-        self.special_tokens_dict = special_tokens_dict
         self._max_possible_token_id = max_possible_token_id
         self._max_special_token_id = max_special_token_id
 
         if not load_adjusted_jsons:
             # store special tokens in a list to preserve their order:
-            all_special_tokens: List
-            if self.special_tokens_dict is None:
-                all_special_tokens = list([])
-            else:
-                all_special_tokens = list(self.special_tokens_dict.values())
+            all_special_tokens = (
+                list(special_tokens_list) if special_tokens_list is not None else []
+            )
 
             # collect all special tokens (without indices):
             for t_type in self.tokenizers_info:
@@ -133,9 +129,9 @@ class ModularTokenizer(transformers.PreTrainedTokenizerFast):
                     )
                 next_index = self._max_special_token_id + 1
         else:
-            if special_tokens_dict is not None:
+            if special_tokens_list is not None:
                 raise Exception(
-                    "When loading a tokenizer special_tokens_dict must be None. Use ModularTokenizer.add_special_tokens instead"
+                    "When loading a tokenizer special_tokens_list must be None. Use ModularTokenizer.add_special_tokens instead"
                 )
             for t_type in self.tokenizers_info:
                 t_info = self.tokenizers_info[t_type]
@@ -161,11 +157,9 @@ class ModularTokenizer(transformers.PreTrainedTokenizerFast):
             # operations on the tokenizer instance (if possible, operations should be done here, using built-in tokenizer methods)
             json_str = json.dumps(t_json)
             tokenizer_inst = Tokenizer.from_str(json_str)
-            if self.special_tokens_dict is not None:
-                # At this point, tokens from self.special_tokens_dict are in every tokenizer. This is only to test that all special tokens were added.
-                num_add = tokenizer_inst.add_special_tokens(
-                    list(self.special_tokens_dict.values())
-                )
+            if special_tokens_list is not None:
+                # At this point, tokens from special_tokens_list are in every tokenizer. This is only to test that all special tokens were added.
+                num_add = tokenizer_inst.add_special_tokens(special_tokens_list)
                 if num_add > 0:
                     raise Exception(
                         f"All special tokens should have been in the vocabulary at this point. {num_add} were added - need to check why."
@@ -1424,15 +1418,7 @@ class ModularTokenizer(transformers.PreTrainedTokenizerFast):
             # operations on the tokenizer instance (if possible, operations should be done here, using built-in tokenizer methods)
             json_str = json.dumps(t_json)
             tokenizer_inst = Tokenizer.from_str(json_str)
-            if self.special_tokens_dict is not None:
-                # At this point, tokens from self.special_tokens_dict are in every tokenizer. This takes care that the special tokens are added to the tokenizer instance.
-                num_add = tokenizer_inst.add_special_tokens(
-                    list(self.special_tokens_dict.values())
-                )
-                if num_add > 0:
-                    raise Exception(
-                        f"All special tokens should have been in the vocabulary at this point. {num_add} were added - need to check why."
-                    )
+
             # restore truncation that was lost when we reset the tokenizer instance
             if "max_len" in t_info and t_info["max_len"] is not None:
                 max_size = t_info["max_len"]
