+++ fusedrug/data/tokenizer/ops/__init__.py
@@ -1,0 +1,8 @@
+# keeping name for backward compatibility
+from fusedrug.data.tokenizer.ops.tokenizer_op import TokenizerOp as FastTokenizer
+from fusedrug.data.tokenizer.ops.modular_tokenizer_op import (
+    ModularTokenizerWithoutInjectOp as FastModularTokenizer,
+)
+from fusedrug.data.tokenizer.ops.modular_tokenizer_op import (
+    ModularTokenizerOp as InjectorTokenizerOp,
+)
